use_cases/pytorch/code/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
Cython==3.0.2
faster-fifo==1.4.5
idna==3.4
numpy==1.26.0
orjson==3.9.7
pandas==2.1.1
Pillow==9.5.0
platformdirs==3.10.0
protobuf==4.24.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
types-requests==2.31.0.3
types-urllib3==1.26.25.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
whylabs-client==0.5.7
whylogs==1.3.4.dev3
whylogs-sketching==3.4.1.dev3
